Moved files and added Docker.
