+++ Project/Backend/mailing-service/docker-compose.yml
@@ -1,5 +1,3 @@
-version: '3.8'
-
 services:
   db:
     image: postgres:15-alpine
@@ -10,8 +8,8 @@ services:
 
     # Env variables, must be unique
     environment:
-      POSTGRES_USER: ${MAILING_DB_USERNAME}
-      POSTGRES_PASSWORD: ${MAILING_DB_PASSWORD}
+      POSTGRES_USER: ${DB_USERNAME}
+      POSTGRES_PASSWORD: ${DB_PASSWORD}
       POSTGRES_DB: ${MAILING_DB_NAME}
 
     volumes:
@@ -51,7 +49,7 @@ services:
     restart: always
 
     # Must be unique
-    container_name: ${PGADMIN_CONTAINER}
+    container_name: ${MAILING_PGADMIN_CONTAINER}
 
     environment:
       PGADMIN_DEFAULT_EMAIL: ${PGADMIN_EMAIL}
