+++ Project/Backend/mailing-service/package.json
@@ -20,5 +20,10 @@
   "devDependencies": {
     "@nestjs/cli": "^11.0.21",
     "@nestjs/schematics": "^11.1.0"
+  },
+  "dependencies": {
+    "@nestjs/common": "^11.1.19",
+    "@nestjs/config": "^4.0.4",
+    "@nestjs/typeorm": "^11.0.1"
   }
 }
