+++ Project/Backend/mailing-service/pnpm-workspace.yaml
@@ -1,0 +1,2 @@
+allowBuilds:
+  '@nestjs/core': set this to true or false
